requirements.txt
aioboto3==11.3.0
aiobotocore==2.6.0
aiofiles==23.1.0
aiohttp==3.8.4
aioitertools==0.11.0
aiosignal==1.3.1
altair==5.0.1
annotated-types==0.5.0
anyio==3.7.1
appnope==0.1.3
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.2
aws-cdk-lib==2.90.0
aws-cdk.asset-awscli-v1==2.2.200
aws-cdk.asset-kubectl-v20==2.1.2
aws-cdk.asset-node-proxy-agent-v5==2.0.166
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
bokeh==3.2.2
boto3==1.28.17
botocore==1.31.17
bsdiff4==1.2.2
cachetools==5.3.1
cattrs==23.1.2
certifi==2022.12.7
cffi==1.15.1
charset-normalizer==3.1.0
click==8.1.3
comm==0.1.4
constructs==10.2.69
contourpy==1.1.0
cryptography==41.0.4
cycler==0.11.0
dataclasses-json==0.5.14
debugpy==1.6.7.post1
decorator==5.1.1
deeplake==3.7.2
dill==0.3.7
distlib==0.3.7
entrypoints==0.4
executing==1.2.0
faiss-cpu==1.7.4
fastapi==0.100.1
ffmpy==0.3.1
filelock==3.12.2
Flask==2.3.1
fonttools==4.41.0
frozenlist==1.3.3
fsspec==2023.4.0
gitdb==4.0.10
GitPython==3.1.37
google-ai-generativelanguage==0.2.0
google-api-core==2.11.1
google-auth==2.20.0
google-generativeai==0.1.0
googleapis-common-protos==1.59.1
gradio==3.39.0
gradio_client==0.3.0
greenlet==2.0.2
grpcio==1.56.0
grpcio-status==1.56.0
h11==0.14.0
httpcore==0.17.3
httpx==0.24.1
huggingface-hub==0.14.1
humbug==0.3.2
idna==3.4
importlib-metadata==6.8.0
importlib-resources==6.0.1
ipykernel==6.25.1
ipython==8.14.0
itsdangerous==2.1.2
jedi==0.18.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.2.0
jsii==1.86.1
jsonschema==4.18.4
jsonschema-specifications==2023.7.1
jupyter_client==8.3.0
jupyter_core==5.3.1
keybert==0.7.0
kiwisolver==1.4.4
langchain==0.0.292
langsmith==0.0.37
libdeeplake==0.0.1
linkify-it-py==2.0.2
markdown-it-py==2.2.0
MarkupSafe==2.1.2
marshmallow==3.20.1
matplotlib==3.7.2
matplotlib-inline==0.1.6
mdit-py-plugins==0.3.3
mdurl==0.1.2
mplfinance==0.12.10b0
mpmath==1.3.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nest-asyncio==1.5.7
networkx==3.1
nltk==3.8.1
numcodecs==0.11.0
numexpr==2.8.6
numpy==1.24.2
openai==0.27.0
orjson==3.9.2
packaging==23.1
pandas==2.0.0
parso==0.8.3
pathos==0.3.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.9.1
pox==0.3.3
ppft==1.7.6.7
prompt-toolkit==3.0.39
proto-plus==1.22.3
protobuf==4.23.4
psutil==5.9.5
ptyprocess==0.7.0
publication==0.0.3
pure-eval==0.2.2
pyarrow==13.0.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycodestyle==2.10.0
pycparser==2.21
pydantic==2.1.1
pydantic_core==2.4.0
pydeck==0.8.1b0
pydub==0.25.1
Pygments==2.15.1
PyJWT==2.8.0
pyparsing==3.0.9
pypdf==3.16.0
PySide6==6.5.1.1
PySide6-Addons==6.5.1.1
PySide6-Essentials==6.5.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pytz==2023.3
PyYAML==6.0
pyzmq==25.1.1
referencing==0.30.0
regex==2023.3.23
requests==2.28.2
requests-file==1.5.1
rich==13.3.4
rpds-py==0.9.2
rsa==4.9
s3transfer==0.6.2
scikit-learn==1.2.2
scipy==1.10.1
semantic-version==2.10.0
sentence-transformers==2.2.2
sentencepiece==0.1.98
shiboken6==6.5.1.1
six==1.16.0
smmap==5.0.1
sniffio==1.3.0
soupsieve==2.4
SQLAlchemy==2.0.20
st-chat-message==0.3.8
stack-data==0.6.2
starlette==0.27.0
streamlit==1.27.0
streamlit-bokeh-events==0.1.2
streamlit-chat==0.1.1
sympy==1.11.1
tenacity==8.2.3
threadpoolctl==3.1.0
tldextract==3.4.0
tokenizers==0.13.3
toml==0.10.2
toolz==0.12.0
torch==2.0.0
torchvision==0.15.1
tornado==6.3.3
tqdm==4.65.0
traitlets==5.9.0
transformers==4.28.1
typeguard==2.13.3
typing-inspect==0.9.0
typing_extensions==4.7.1
tzdata==2023.3
tzlocal==5.0.1
uc-micro-py==1.0.2
urllib3==1.26.17
uvicorn==0.23.1
validators==0.22.0
virtualenv==20.24.2
watchdog==3.0.0
wcwidth==0.2.6
websockets==11.0.3
Werkzeug==2.3.2
wrapt==1.15.0
xyzservices==2023.7.0
yarl==1.8.2
zipp==3.17.0
